chore: check diff support args
