--- bin/antd-tools-run.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require('../lib/cli/run');
