+++ lib/cli/index.js
@@ -2,26 +2,40 @@
 
 'use strict';
 
-require('colorful').colorful();
+const chalk = require('chalk');
+const gulp = require('gulp');
 
-const program = require('commander');
-const packageInfo = require('../../package.json');
+const argv = require('minimist')(process.argv.slice(2));
 
-program
-  .version(packageInfo.version)
-  .command('run [name]', 'run specified task')
-  .parse(process.argv);
+const cloneArgs = { ...argv };
+delete cloneArgs._;
 
-// https://github.com/tj/commander.js/pull/260
-const proc = program.runningCommand;
-if (proc) {
-  proc.on('close', process.exit.bind(process));
-  proc.on('error', () => {
-    process.exit(1);
-  });
+console.log(chalk.yellow('Execute:'), chalk.green(argv._[1]), '-', JSON.stringify(cloneArgs));
+console.log('  - Args:', JSON.stringify(cloneArgs));
+
+require('../gulpfile');
+
+// Start glup task
+function runTask(toRun) {
+  const metadata = { task: toRun };
+  // Gulp >= 4.0.0 (doesn't support events)
+  const taskInstance = gulp.task(toRun);
+  if (taskInstance === undefined) {
+    gulp.emit('task_not_found', metadata);
+    return;
+  }
+  const start = process.hrtime();
+  gulp.emit('task_start', metadata);
+  try {
+    taskInstance.apply(gulp);
+    metadata.hrDuration = process.hrtime(start);
+    gulp.emit('task_stop', metadata);
+    gulp.emit('stop');
+  } catch (err) {
+    err.hrDuration = process.hrtime(start);
+    err.task = metadata.task;
+    gulp.emit('task_err', err);
+  }
 }
 
-const subCmd = program.args[0];
-if (!subCmd || subCmd !== 'run') {
-  program.help();
-}
+runTask(argv._[1]);
