+++ lib/lint/checkDiff.js
@@ -1,11 +1,16 @@
 const { getProjectPath } = require('../utils/projectHelper'); // eslint-disable-line import/order
 
 const fs = require('fs');
+const { join } = require('path');
 const chalk = require('chalk');
 const fetch = require('node-fetch');
 const readline = require('readline');
 
-function getMajorVersion(version) {
+function getMajorVersion(version, specificVersion) {
+  if (specificVersion) {
+    return `@${specificVersion}`;
+  }
+
   const match = version && version.match(/^\d+/);
 
   if (match) {
@@ -20,9 +25,15 @@ function getVersionFromURL(url, name) {
   return affix.slice(0, affix.indexOf('/'));
 }
 
+const argv = require('minimist')(process.argv.slice(2));
+
 module.exports = function (packageName, packageVersion, done) {
-  console.log(chalk.cyan('Fetching latest version file list...'));
-  fetch(`https://unpkg.com/${packageName}${getMajorVersion(packageVersion)}/?meta`)
+  console.log('>>>', process.argv, argv);
+
+  const mergedVersion = getMajorVersion(packageVersion, argv.version);
+  console.log(chalk.cyan(`Fetching latest version file list...${packageName}${mergedVersion}`));
+
+  fetch(`https://unpkg.com/${packageName}${mergedVersion}/?meta`)
     .then(res => {
       const version = getVersionFromURL(res.url, packageName);
       return res.json().then(json => ({ version, ...json }));
@@ -40,7 +51,11 @@ module.exports = function (packageName, packageVersion, done) {
     })
     .then(({ version, fileList }) => {
       // Find missing files
-      const missingFiles = fileList.filter(filePath => !fs.existsSync(getProjectPath(filePath)));
+      const missingFiles = fileList.filter(filePath => {
+        const concatFilePath = argv.path ? join(argv.path, filePath) : filePath;
+
+        return !fs.existsSync(getProjectPath(concatFilePath));
+      });
 
       if (missingFiles.length) {
         console.log(
